feat: extract repeated code into components
