+++ src/css/whats-new.css
@@ -1,14 +1,6 @@
 * {
-  font-family: 'Roboto', system-ui, sans-serif;
   font-size: 14px;
   box-sizing: border-box;
-}
-
-body {
-  padding: 0;
-  margin: 0;
-  background-color: #2d2d2d;
-  color: #fff;
 }
 
 .whats-new-root {
@@ -17,27 +9,26 @@ body {
   padding: 20px;
 }
 
-.whats-new-header {
-  text-align: center;
-  background-color: #3d3d3d;
+/* Shared component overrides for wider tab layout */
+.whats-new-root .header {
   padding: 20px;
   margin-bottom: 20px;
 }
 
-.whats-new-logo {
+.whats-new-root .header .logo {
   width: 60px;
 }
 
-.whats-new-title {
+.whats-new-root .header .title {
   font-size: 24px;
   margin-top: 10px;
-  margin-bottom: 0;
 }
 
-.whats-new-explanation {
-  background-color: #3d3d3d;
+.whats-new-root .footer {
+  margin-top: 20px;
 }
 
+/* Changelog content */
 .whats-new-version {
   background-color: #3d3d3d;
   padding: 15px 20px;
@@ -83,90 +74,4 @@ body {
 .whats-new-list-item {
   margin-bottom: 4px;
   line-height: 1.5;
-}
-
-:root {
-  --animation-speed: 0.2s;
-}
-
-a {
-  color: #fff;
-  text-decoration: none;
-}
-
-.button {
-  color: #fff;
-  text-decoration: none;
-  cursor: pointer;
-  background-color: #3d3d3d;
-  padding: 10px;
-  border: 1px solid #3d3d3d;
-  text-align: center;
-  display: inline-block;
-  width: 100%;
-  transition: all var(--animation-speed);
-}
-
-.button:not(:disabled):hover {
-  background-color: #4d4d4d;
-}
-
-.bold {
-  font-weight: bold;
-}
-
-.small {
-  font-size: 9px;
-}
-
-.name {
-  font-weight: bold;
-  font-size: inherit;
-}
-
-.whats-new-footer {
-  margin-top: 20px;
-}
-
-.credits {
-  padding: 10px;
-  text-align: center;
-  background-color: #3d3d3d;
-  display: block;
-  transition: all var(--animation-speed);
-}
-
-.credits.small {
-  font-size: 9px;
-}
-
-.bmac {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  row-gap: 8px;
-  padding: 8px;
-  background: linear-gradient(
-    to top left,
-    rgba(255, 0, 0, 0.3) 10%,
-    transparent
-  );
-  opacity: 0.5;
-  transition: all var(--animation-speed) ease-in-out;
-}
-
-.bmac * {
-  font-size: inherit;
-}
-
-.bmac:hover {
-  opacity: 1;
-}
-
-.bmac-link {
-  transition: all var(--animation-speed) ease-in-out;
-}
-
-.bmac-link:hover {
-  opacity: 0.75;
 }
